feat: enhance folder import modal with progress UI
